% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat1.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Sample from a normal distribution}
\format{
\subsection{\code{vector}}{

A vector with 10000 samples:
\describe{
\item{sample}
...
}
}
}
\usage{
dat1
}
\description{
Sample of size 10000 from a normal distribution of mean 10 and sd 2
See...
}
\keyword{datasets}
