% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inhouse_data}
\alias{inhouse_data}
\title{World Health Organization TB data}
\format{
\subsection{\code{tibble}}{

A tibble with 100 rows and 50 columns:
\describe{
\item{customerq1}{s1}
\item{customerq1}{s2}
\item{customerq3}{s3}
\item{customerq4}{s4}
\item{customerq5}{s5}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
inhouse_data
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
